[FE2#1,2,3] feat&refactor : 권한 띄우기, 시작 페이지(키패드), 설정 페이지
