+++ pubspec.yaml
@@ -32,6 +32,7 @@ dependencies:
     sdk: flutter
   permission_handler : ^10.2.0
   url_launcher : ^6.0.20
+  shared_preferences : ^2.0.6
 
 
   # The following adds the Cupertino Icons font to your application.
